+++ Clicked/Core/Database.lua
@@ -235,30 +235,6 @@ function Clicked:GetBindingIndex(binding)
 	return 0
 end
 
---- Get the active action of a binding configuration. The data for spells, items,
---- and macros is all saved in separate data structures. This function will return
---- the correct data structure for the current `type` of the binding.
----
---- @param binding table
---- @return table
-function Clicked:GetActiveBindingValue(binding)
-	assert(type(binding) == "table", "bad argument #1, expected table but got " .. type(binding))
-
-	if binding.type == Clicked.BindingTypes.SPELL then
-		return binding.action.spellValue
-	end
-
-	if binding.type == Clicked.BindingTypes.ITEM then
-		return binding.action.itemValue
-	end
-
-	if binding.type == Clicked.BindingTypes.MACRO then
-		return binding.action.macroValue
-	end
-
-	return nil
-end
-
 function Clicked:GetBindingCache(binding)
 	return binding.cache
 end
@@ -727,6 +703,11 @@ function Clicked:UpgradeDatabaseProfile(profile, from)
 	-- 0.15.x to 0.16.0
 	if string.sub(from, 1, 4) == "0.15" then
 		profile.options.tooltips = false
+
+		for _, binding in ipairs(profile.bindings) do
+			binding.integrations = {}
+		end
+
 		FinalizeVersionUpgrade("0.16.0")
 	end